+++ js/main.js
@@ -1,11 +1,4 @@
-import {
-  createAd,
-  ADS_COUNT
-} from './data.js';
-
-import {
-  createCard
-} from './card.js';
+import { createAds } from './data.js';
 
 import {
   activateMapFilters,
@@ -17,20 +10,10 @@ import {
   deactivateForm
 } from './form.js';
 
-import {
-  addMap,
-  addMarkers
-} from './map.js';
+import { addMap } from './map.js';
 
 // Создание данных объявлений
-const createAds = (count) => {
-  const ads = [];
-  while (ads.length < count) {
-    ads.push(createAd());
-  }
-  return ads;
-};
-const adsData = createAds(ADS_COUNT);
+const adsData = createAds();
 
 // Функции активации страницы
 const activatePage = () => {
@@ -43,15 +26,7 @@ const deactivatePage = () => {
   deactivateForm();
 };
 
+deactivatePage();
+
 // Управление картой
-const loadMap = () => {
-  const map = addMap(activatePage, deactivatePage);
-
-  adsData.forEach((ad) => {
-    const location = ad.location;
-    const card = createCard(ad);
-    addMarkers(map, location, card);
-  });
-};
-
-loadMap();
+addMap(adsData, activatePage);
